rename: 파일명 간단화
